remove: vercel.json 삭제
